+++ backtrader/plot/bokeh/__init__.py
@@ -1,4 +1,4 @@
-from .bokeh.bokeh import Bokeh
+from .bokeh import Bokeh
 
 # initialize analyzer tables
 from .analyzer_tables import inject_datatables
