+++ backtrader/plot/bokeh/analyzer_tables/calmar.py
@@ -1,8 +1,8 @@
-from ..bokeh.datatable import ColummDataType
+from ..datatable import ColummDataType
 
 
 def datatable(self):
-    cols1 = [['DateTime', ColummDataType.DATETIME], ['Return', ColummDataType.FLOAT]]
+    cols1 = [['DateTime', ColummDataType.DATETIME], ['Value', ColummDataType.FLOAT]]
 
     a = self.get_analysis()
 
@@ -10,9 +10,8 @@ def datatable(self):
         cols1[0].append(k)
         cols1[1].append(v)
 
-    name = "Time Return"
+    name = "Calmar"
     if self.p.plotname:
         name = self.p.plotname
 
     return name, [cols1]
-
