+++ backtrader/plot/bokeh/analyzer_tables/leverage.py
@@ -1,8 +1,8 @@
-from ..bokeh.datatable import ColummDataType
+from ..datatable import ColummDataType
 
 
 def datatable(self):
-    cols1 = [['DateTime', ColummDataType.DATETIME], ['Value', ColummDataType.FLOAT]]
+    cols1 = [['DateTime', ColummDataType.DATETIME], ['Leverage', ColummDataType.FLOAT]]
 
     a = self.get_analysis()
 
@@ -10,7 +10,7 @@ def datatable(self):
         cols1[0].append(k)
         cols1[1].append(v)
 
-    name = "Calmar"
+    name = "Gross Leverage"
     if self.p.plotname:
         name = self.p.plotname
