+++ backtrader/plot/bokeh/analyzer_tables/sqn.py
@@ -1,14 +1,17 @@
-from ..bokeh.datatable import ColummDataType
+from ..datatable import ColummDataType
 
 
 def datatable(self):
     cols1 = [['Name', ColummDataType.STRING], ['Value', ColummDataType.FLOAT]]
 
     a = self.get_analysis()
-    cols1[0].append('VWR')
-    cols1[1].append(a['vwr'])
+    cols1[0].append('SystemQualityNumber')
+    cols1[1].append(a['sqn'])
 
-    name = "Variability-Weighted Return"
+    cols1[0].append('Trades')
+    cols1[1].append(a['trades'])
+
+    name = "System Quality Number"
     if self.p.plotname:
         name = self.p.plotname
