+++ backtrader/plot/bokeh/analyzer_tables/timereturn.py
@@ -1,8 +1,8 @@
-from ..bokeh.datatable import ColummDataType
+from ..datatable import ColummDataType
 
 
 def datatable(self):
-    cols1 = [['DateTime', ColummDataType.DATETIME], ['Leverage', ColummDataType.FLOAT]]
+    cols1 = [['DateTime', ColummDataType.DATETIME], ['Return', ColummDataType.FLOAT]]
 
     a = self.get_analysis()
 
@@ -10,8 +10,9 @@ def datatable(self):
         cols1[0].append(k)
         cols1[1].append(v)
 
-    name = "Gross Leverage"
+    name = "Time Return"
     if self.p.plotname:
         name = self.p.plotname
 
     return name, [cols1]
+
