+++ backtrader/plot/bokeh/schemes/scheme.py
