+++ backtrader/plot/bokeh/schemes/tradimo.py
