+++ backtrader/plot/matplot/utils.py
